re-configured database access
Now done with context.xml and web.xml instead of using a
ServletContextListener to store a POJ DataSource in the context.

Site is broken with this build as I am mid-way through adding DAO access
to Models and other service layer object.

Also contains error with no hasRole() method being available in some of
the User classes.